Fixes mount and snippets

+++ .devcontainer/postCreateScript.sh
@@ -47,20 +47,23 @@ readonly SNIPPET_FILES_PATTERN="*.code-snippets"
 curl -fsSL https://raw.githubusercontent.com/gvatsal60/Linux-Aliases/HEAD/install.sh | sh
 
 
-# Generate symlinks for snippet files
-# Create the .vscode directory if it doesn't already exist.
-mkdir -p "${VSCODE_DIR}"
+# Check if running in GitHub Codespaces by checking for `CODESPACE_NAME` env variable
+if [ -n "${CODESPACE_NAME}" ]; then
+    # Generate symlinks for snippet files
+    # Create the .vscode directory if it doesn't already exist.
+    mkdir -p "${VSCODE_DIR}"
 
-# Unlink all symbolic links in the '.vscode' directory
-find "${VSCODE_DIR}" -type l -name "${SNIPPET_FILES_PATTERN}" | while read -r file; do
-    unlink "${file}"
-done
-
-# Check if the 'snippets' directory exists
-if [ -d "${SNIPPETS_DIR}" ]; then
-    # Find all .code-snippet files in the 'snippets' directory and create symbolic links
-    find "${SNIPPETS_DIR}" -type f -name "${SNIPPET_FILES_PATTERN}" | while read -r file; do
-        # Create a symbolic link in '.vscode' with the same base name
-        ln -s "$(realpath "${file}")" "${VSCODE_DIR}/$(basename "${file}")"
+    # Unlink all symbolic links in the '.vscode' directory
+    find "${VSCODE_DIR}" -type l -name "${SNIPPET_FILES_PATTERN}" | while read -r file; do
+        unlink "${file}"
     done
+
+    # Check if the 'snippets' directory exists
+    if [ -d "${SNIPPETS_DIR}" ]; then
+        # Find all .code-snippet files in the 'snippets' directory and create symbolic links
+        find "${SNIPPETS_DIR}" -type f -name "${SNIPPET_FILES_PATTERN}" | while read -r file; do
+            # Create a symbolic link in '.vscode' with the same base name
+            ln -s "$(realpath "${file}")" "${VSCODE_DIR}/$(basename "${file}")"
+        done
+    fi
 fi
